--- src/main/java/searchengine/services/LemmaService.java
@@ -1,8 +1,0 @@
-package searchengine.services;
-
-import searchengine.model.LemmaEntity;
-import searchengine.model.SiteEntity;
-
-public interface LemmaService {
-    LemmaEntity saveOrUpdateLemma(String lemma, SiteEntity site);
-}
