+++ src/main/java/searchengine/services/SearchService.java
@@ -3,7 +3,5 @@ package searchengine.services;
 import searchengine.dto.search.SearchResponse;
 
 public interface SearchService {
-    SearchResponse search(String query, String siteUrl, int offset, int limit);
-
-    SearchResponse searchAll(String query, int offset, int limit);
+    SearchResponse search(String query, String site, int offset, int limit);
 }
